requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
mimerender==0.6.0
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pytest==3.2.3
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.15
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.12.2
